Linked list
While processing if the formula has a valid format, we will be storing
each separated register (between ',') into separated elements on a
linked list.

The list will be firstly processed by  '*' and '/' operations and then
by '+' and '-'.

Note: We haven´t freed the previuous nodes yet so it has to be done soon